pyproject.toml
[tool.poetry]
name = "lapi"
version = "0.1.0"
description = "Lamden API"
authors = ["Endogen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pynacl = "^1.5.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
